% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyAlns.R
\name{modifyAlns}
\alias{modifyAlns}
\title{modifyAlns.R}
\usage{
modifyAlns(reference, compare, filename)
}
\arguments{
\item{reference}{A directory path to the reference alignment
the user would like to use}

\item{compare}{A directory path to the alignment being compared
to the reference alignment}

\item{filename}{The preferred name of the output file provided by
the user in ScoreVisualizer}
}
\value{
The filepaths of the copied and modified alignments that
will be used to generate the tex file
}
\description{
The \code{modifyAlns} function copies and modifies the 
alignments provided by the user in order to generate a valid
texShade output. It adds a 'sample' alignment since 
texShade does not include gaps in the sequence 
length of the alignment. TexShare only counts the 
residues in the individual sequences. Lastly, the
function returns the filepaths of the copied and modified 
alignments.

Details.
}
\examples{
reference <- "data/BB11017ClustalW.fasta"
test <- "data/BB11017Muscle.fasta"

filepaths <- modifyAlns(reference, compare, "Test3")
filepaths

}
\keyword{internal}
