% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreVisualizer.R
\name{ScoreVisualizer}
\alias{ScoreVisualizer}
\title{ScoreVisualizer.R}
\usage{
ScoreVisualizer(reference, compare, SP = TRUE, CS = TRUE, fileName = NULL)
}
\arguments{
\item{reference}{A directory path to the reference alignment
the user would like to use}

\item{compare}{A directory path to the alignment being compared
to the reference alignment}

\item{SP}{A boolean value left TRUE if the user would like to
generate the sum of pairs in their output}

\item{CS}{A boolean value left TRUE if the user would like
to generate the total column score in their output}

\item{fileName}{The preferred name of the output file}
}
\description{
The \code{ScoreVisualizer} function is the main function
for the ScoreVisualizer package. It calls the helper functions
and generates a visual output to show the degree of scoring
within two alignments. 

Details.
}
\examples{
reference <- "data/BB11017ClustalW.fasta"
test <- "data/BB11017Muscle.fasta"
ScoreVisualizer(reference, test)

}
