% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFileExists.R
\name{checkFileExist}
\alias{checkFileExist}
\title{checkFileExist}
\usage{
checkFileExist(file)
}
\arguments{
\item{file}{The filepath of the file of interest in the form
of a character vector.}
}
\description{
The \code{checkFileExist} function checks to see if
the filename requested is already existing in the directory. If
it is, prompt the user if they would like to overwrite. If no,
then stop the running functions.

Details.
}
\keyword{internal}
