% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFiles.R
\name{deleteFiles}
\alias{deleteFiles}
\title{deleteFiles.R}
\usage{
deleteFiles(filePaths)
}
\arguments{
\item{filePaths}{A vector of filePaths directing to files that
are needed to be removed}
}
\description{
The \code{deleteFiles} function removes any existing
files given the filepath

Details.
}
