% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{getScore}
\alias{getScore}
\title{getScore}
\usage{
getScore(reference, compare, SP = TRUE, CS = TRUE)
}
\arguments{
\item{reference}{A directory path to the reference alignment
the user would like to use}

\item{compare}{A directory path to the alignment being compared
to the reference alignment}

\item{SP}{A boolean value left TRUE if the user would like to
generate the sum of pairs in their output}

\item{CS}{A boolean value left TRUE if the user would like
to generate the total column score in their output}
}
\value{
The Pairwise alignment comparison object from the two
alignments provided
}
\description{
The \code{getScore} function generates a "Pairwise alignment
comparison" (PAC) object which contains the sum of pairs
score and the total column score of two multiple sequence
alignments.

Details.
}
\examples{
score <- getScore("data/sampleref.fasta", "data/sampleMuscle.fasta")
score

}
