% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScoreFile.R
\name{makeScoreFile}
\alias{makeScoreFile}
\title{makeScoreFile}
\usage{
makeScoreFile(reference, compare, SP = TRUE, CS = TRUE, filename, score)
}
\arguments{
\item{reference}{A directory path to the reference alignment
the user would like to use}

\item{compare}{A directory path to the alignment being compared
to the reference alignment}

\item{SP}{A boolean value left TRUE if the user would like to
generate the sum of pairs in their output}

\item{CS}{A boolean value left TRUE if the user would like
to generate the total column score in their output}

\item{file}{The preferred name of the output file}
}
\description{
The \code{makeScoreFile} function generates score files
containing columnwise sum of pairs score and total column
score. 

Details.
}
